+++ examples/StateComponents/transitionOnCondition.ts
@@ -1,21 +1,21 @@
 import {
     IMachineSPI,
+    StateSetupFn,
 } from '../..';
 
 /**
- *
+ * Immediately transitions to one of two states based on the result of the
+ * condition function.
  */
 export const transitionOnCondition = <T extends string>(
     condition: () => boolean,
     successState: T,
     failState: T,
-) => {
-    return () => {
-        return {
-            onRun: (machine: IMachineSPI<T, any>) => machine.transitionToState(condition()
-                ? successState
-                : failState
-            ),
-        };
+): StateSetupFn<T, any> => {
+    return (machine: IMachineSPI<T, any>) => {
+        machine.transitionToState(condition()
+            ? successState
+            : failState
+        );
     }
 }