+++ src/stateComponents/index.ts
@@ -1,3 +1,2 @@
 export { onEnter } from './onEnter';
 export { onExit } from './onExit';
-export { onRun } from './onRun';