+++ src/types/StateRunFn.ts
@@ -1,5 +1,0 @@
-import { IMachineSPI } from '../IMachineSPI';
-
-export type StateRunFn<T extends string, D> = {
-    (machine: IMachineSPI<T, D>): void,
-};