+++ src/types/index.ts
@@ -1,5 +1,4 @@
 export { MachineBlueprint } from './MachineBlueprint';
 export { StateExitFn } from './StateExitFn';
-export { StateRunFn } from './StateRunFn';
 export { StateSetupFn } from './StateSetupFn';
 export { MachineTerminateFn } from './MachineTerminateFn';