+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 This project is currently in design phase, and as such does not follow Semantic Versioning. However, it will adopt Semantic Versioning beginning with version 1.0.0.
 <!-- and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html). -->
 
+## Unreleased
+
+### Added
+- Can now trigger transitions from a StateComponent's setup function.
+
+### Changed
+- The `StateSetupFn` now returns a single optional cleanup function, rather than `{ onRun, onExit }`.
+
+### Removed
+- Type `StateRunFn`.
+- The `onRun` StateComponent.
+
 ## 0.4.2
 
 ### Changed